Move ChainReadError and ChainReader to libsawtooth
